Move IndexTestHelpers.h into correct folder
