+++ test/MultiColumnJoinTest.cpp
@@ -7,12 +7,12 @@
 #include <array>
 #include <vector>
 
-#include "./IndexTestHelpers.h"
 #include "./util/AllocatorTestHelpers.h"
 #include "./util/IdTableHelpers.h"
 #include "./util/IdTestHelpers.h"
 #include "engine/CallFixedSize.h"
 #include "engine/MultiColumnJoin.h"
+#include "util/IndexTestHelpers.h"
 
 using ad_utility::testing::makeAllocator;
 namespace {
