+++ test/OrderByTest.cpp
@@ -5,12 +5,12 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include "./IndexTestHelpers.h"
 #include "./util/IdTableHelpers.h"
 #include "./util/IdTestHelpers.h"
 #include "engine/OrderBy.h"
 #include "engine/ValuesForTesting.h"
 #include "global/ValueIdComparators.h"
+#include "util/IndexTestHelpers.h"
 
 using namespace std::string_literals;
 using namespace std::chrono_literals;
